backend/denoising-ai/requirements_dev.txt
# Test
coverage==5.5
httpx>=0.23,<0.27
moto==2.2.13
mock==4.0.3
pylint==2.11.1
pytest==6.2.4

# API / FastAPI stack
fastapi==0.109.0
uvicorn==0.35.0
python-multipart==0.0.20
starlette-context==0.3.6
starlette==0.35.1

# Pydantic
pydantic==2.11.9
pydantic-settings==2.10.1

# Procesamiento de imágenes
numpy==1.24.3
Pillow==11.3.0
pywavelets==1.8.0

# Deep learning / modelos
torch==2.8.0
torchvision==0.23.0
deepinv==0.3.4

# AWS etc
boto3==1.40.30
botocore==1.40.30
aws-secretsmanager-caching==1.1.3

# Utilidades
requests==2.28.2
#structlog==21.1.0
structlog==25.4.0
protobuf==4.25.8